# Conflicts:
#	src/game/GameController.java

+++ src/util/Math.java
@@ -14,7 +14,7 @@ public class Math {
      * @param max maximum number to return
      * @return an integer inside the domain
      */
-    public static int clamp(int num, int min, int max) {
+    public static double clamp(double num, double min, double max) {
         if (num > max) {
             return max;
         }
@@ -32,7 +32,7 @@ public class Math {
      * @param amount how much to approach
      * @return integer of the maximum approach it could do
      */
-    public static int approach(int num, int result, double amount) {
+    public static double approach(double num, double result, double amount) {
         if(num < result) {
             num += amount;
             if (num > result)
